Move supertokens config into lib
