+++ api/src/functions/auth.js
@@ -1,7 +1,7 @@
 import SuperTokens from 'supertokens-node'
 import { middleware } from 'supertokens-node/framework/awsLambda'
 
-import { config } from '../services/supertokens'
+import { config } from '../lib/supertokens'
 
 SuperTokens.init(config)
